Implementing various tuple features

+++ src/main/java/org/aicl/raytracerchallenge/geometry/Point.java
@@ -1,7 +1,7 @@
 package org.aicl.raytracerchallenge.geometry;
 
 public class Point extends Tuple {
-    public Point(float x, float y, float z){
+    public Point(double x, double y, double z){
         super(x, y, z, 1);
     }
 }
