+++ src/main/java/org/aicl/raytracerchallenge/utilities/FloatEquality.java
@@ -3,7 +3,7 @@ package org.aicl.raytracerchallenge.utilities;
 import org.aicl.raytracerchallenge.geometry.Constant;
 
 public class FloatEquality {
-    public static boolean isEqual(float a, float b){
+    public static boolean isEqual(double a, double b){
         return Math.abs(a - b) < Constant.epsilon;
     }
 }
